# Optimize writing optional columns (part 1)