This PR is the first part of a refactor intended to optimize the encoding of optional columns.

I modified the `ColumnBuffer` and `Dictionary` interface by adding an unexported method which replaces the type switch dispatch we were previously doing, and both enables the optimizations on the optional column and greatly reduced code duplication; @kevinburkesegment I believe this should address the comment you made in https://github.com/segmentio/parquet-go/pull/186#discussion_r880851224

```
name                                          old time/op  new time/op  delta
GenericBuffer/benchmarkRowType/go1.18          751ns ± 0%   751ns ± 0%      ~     (p=0.424 n=10+10)
GenericBuffer/booleanColumn/go1.18             316ns ± 0%   312ns ± 0%    -1.06%  (p=0.000 n=8+9)
GenericBuffer/int32Column/go1.18               236ns ± 0%   230ns ± 0%    -2.64%  (p=0.000 n=10+9)
GenericBuffer/int64Column/go1.18               308ns ± 0%   304ns ± 0%    -1.32%  (p=0.000 n=9+10)
GenericBuffer/floatColumn/go1.18               236ns ± 0%   230ns ± 0%    -2.61%  (p=0.000 n=9+10)
GenericBuffer/doubleColumn/go1.18              308ns ± 0%   304ns ± 0%    -1.29%  (p=0.000 n=9+10)
GenericBuffer/byteArrayColumn/go1.18          9.90µs ± 1%  9.91µs ± 1%      ~     (p=0.400 n=9+10)
GenericBuffer/fixedLenByteArrayColumn/go1.18  5.61µs ± 4%  5.45µs ± 0%    -2.86%  (p=0.000 n=8+10)
GenericBuffer/stringColumn/go1.18             6.21µs ± 0%  6.27µs ± 0%    +0.99%  (p=0.000 n=10+10)
GenericBuffer/indexedStringColumn/go1.18      40.2µs ± 1%  29.2µs ± 0%   -27.45%  (p=0.000 n=9+8)
GenericBuffer/uuidColumn/go1.18                392ns ± 0%   391ns ± 1%      ~     (p=0.138 n=10+10)
GenericBuffer/mapColumn/go1.18                 701µs ± 0%   622µs ± 1%   -11.36%  (p=0.000 n=10+8)
GenericBuffer/decimalColumn/go1.18             308ns ± 0%   303ns ± 0%    -1.46%  (p=0.000 n=10+9)
GenericBuffer/contact/go1.18                  13.0µs ± 0%  13.0µs ± 0%    +0.33%  (p=0.026 n=8+8)
GenericBuffer/paddedBooleanColumn/go1.18       224ns ± 1%   216ns ± 0%    -3.26%  (p=0.000 n=10+10)
GenericBuffer/optionalInt32Column/go1.18      10.8µs ± 0%   2.0µs ± 0%   -81.77%  (p=0.000 n=10+10)

name                                          old row/s    new row/s    delta
GenericBuffer/benchmarkRowType/go1.18          1.33G ± 0%   1.33G ± 0%      ~     (p=0.481 n=10+10)
GenericBuffer/booleanColumn/go1.18             3.17G ± 0%   3.20G ± 0%    +1.07%  (p=0.000 n=8+9)
GenericBuffer/int32Column/go1.18               4.23G ± 0%   4.35G ± 0%    +2.71%  (p=0.000 n=10+9)
GenericBuffer/int64Column/go1.18               3.25G ± 0%   3.29G ± 0%    +1.33%  (p=0.000 n=9+10)
GenericBuffer/floatColumn/go1.18               4.23G ± 0%   4.34G ± 0%    +2.68%  (p=0.000 n=9+10)
GenericBuffer/doubleColumn/go1.18              3.25G ± 0%   3.29G ± 0%    +1.30%  (p=0.000 n=9+10)
GenericBuffer/byteArrayColumn/go1.18            101M ± 1%    101M ± 1%      ~     (p=0.400 n=9+10)
GenericBuffer/fixedLenByteArrayColumn/go1.18    178M ± 4%    183M ± 0%    +2.93%  (p=0.000 n=8+10)
GenericBuffer/stringColumn/go1.18               161M ± 0%    159M ± 0%    -0.98%  (p=0.000 n=10+10)
GenericBuffer/indexedStringColumn/go1.18       24.9M ± 1%   34.3M ± 0%   +37.83%  (p=0.000 n=9+8)
GenericBuffer/uuidColumn/go1.18                2.55G ± 0%   2.56G ± 1%      ~     (p=0.143 n=10+10)
GenericBuffer/mapColumn/go1.18                 1.43M ± 0%   1.61M ± 1%   +12.82%  (p=0.000 n=10+8)
GenericBuffer/decimalColumn/go1.18             3.25G ± 0%   3.30G ± 0%    +1.48%  (p=0.000 n=10+9)
GenericBuffer/contact/go1.18                   77.2M ± 0%   76.9M ± 0%    -0.33%  (p=0.028 n=8+8)
GenericBuffer/paddedBooleanColumn/go1.18       4.47G ± 1%   4.62G ± 0%    +3.36%  (p=0.000 n=10+10)
GenericBuffer/optionalInt32Column/go1.18       92.6M ± 0%  507.6M ± 0%  +448.49%  (p=0.000 n=10+10)
```